Client checker (#12)
Older versions of iOS don't play well with Vuetify, so rather than render a broken app, we check for valid iOS version, and if invalid, we emit "invalid_client" and do not load the first step. This can be overriden by passing setting :validClient=true as a prop in step-list.